First off, let me demonstrate the new patterns.
1) {a=x | s = x}
2) {(a | s) = x}

This is with an arbitrary number of `|`s.

In the () pattern, I might also allow ands.

There are few problems the above opens up. The `=` in the above pattern is a rebind. #1 and #2 are allowed, but what would I do with a pattern like this?

3) {(a=q | s)=x} ...

Well, since {a=q=x} would be illegal, so should the above.

The way I will implement this is by desugaring the {(a | s) = x} into {a=x | s = x}.

First, let me implement it in pattern compile and then I will sketch the thing out. I am not sure even if the way module is structured in that function is for what I want to do. I might need to introduce And  instead of using a list.