It is time for me to start this thing.
I've decided to put in the exclusive or and not into the pattern matcher.

The exclusive or won't be easy and will likely take me the whole day, but let me just get this out of the way.

Something like this will definitely be a great asset in the language, so I should do it.

But before I start on that, let me just deal with the extension patterns. I will change them a bit so I can use the named tuple syntax with them.