Well, it seems all tests pass for the extension pattern change.
Good. Now comes the exclusive or. Let me just extend the module syntax a bit more and then I will be able to start programming for real.

This has been the case for the last few months. I would start doing a problem a problem and then find spots where the language needs to be extended.

This is true programming. If I walk this path, I have no doubt that I will arrive at my destination.