I really don't want to have two alternative structures in the same parser.
No, though the test passes now, I am definitely not done. And in the above thing, the and pattern does not work.

But really, instead of doing one or the other, couldn't I do something like

{((a=t) | b)=t} -> {(a = t) | (b = t)}

I mean does it necessarily have to be an error?

I just had a bit of insight, wouldn't it necessarily be more useful if the rebind isn't distributed where it can't be?

I am putting in a massive amount of effort into essentially nothing.

And is it really true that I can't rebind modules? I mean what is the `and` pattern for?

That way I could also rebind the rebindings.

{a=b=c} -> {a=b a=c}

This is absolute genius. Let me implement it. I am going to have to remove the alt parser I did and redo it, but it won't be too hard.