Finally done with `Save The Princess` problem.
This particular example required me to use 8 different modules and literally pushed me to my limit.

The problem would be better solved with an actual queue rather than maping and array to an array of arrays and concatting it, but for now this is the path of least resistance.

Fuuuuu...I need a breather after all of this.

Hmmhmmm...I actually did not have anything in mind after this since this was such a huge challenge.

I really have to get on with doing full .NET integration, but before that, I want to challenge myself by doing even harder problems in Spiral.

That will be my focus for the next week or so.

What I've done here is tremendous. With every step I take, Spiral comes closer to maturity.

How about I relax for the next week just by doing these problems?

That is what I was planning to do all this time, but I ended up writing libraries and putting new things into the language.

Maybe it is time for this kind of much vaunted 'vacation' to start?