Since it would have made things easier for Save The Princess, and will no doubt be necessary for other dynamic programming kind of problems, here is the queue module.
There is some inefficiencies in making this in Spiral since I am updating all the fields every time, but it should do nicely regardless.

With this, I've done a decent chunk of work, definitely enough to wipe the shame of my earlier laziness.

Tomorrow, I will test the queue and work on more HackerRank of Project Euler problems.

It is time to stop for the day.