Done testing the queue. As expected I was right to be skeptical of its correctness.
The way it is done is pretty brilliant. It is not as simple to get right as it seems on the first glance since the buffer wraps around, but I trully managed to capture the essence of it in just two functions.

Now with that out of the way, what is next? Since Save The Princess 2 is mostly similar to the first one, I think I will skip it. It would just waste space.

Instead let me pick some other problem and deal with it.