Mental state: utter perplexity.
If this is a vacation I want my money back. But yes, the for loops were not made for changing states, though I did expect it to work. Forget the hacker_rank problem for a bit.

Let me separate out the issue and debug it that way. I need some pesticide.