Fixed the bug in `type_annot`.
The compilation bug was a genuine one. It took me a lot of thinking to figure out what it was.

As it turns out, the debugger was useless, my intuition can take all the credit for fixing this here.

But even after I homed in on the right area, it still took me over an hour of though to understand the problem and fix it.

What I will do next is fix the state passing bug and complete hacker_rank_3. After that, maybe I will call it a day.