I had to change the size of the indentation to get around the 50kb limit, but I am done with Save The Princess 2. Also I am done with HackerRank. Not optimizing tail calls and now size limits precludes it from me able to do serious problems on it.
I need a better site.

Well, foget about it for now.

Also I tried to fix the state passing bug in the library, but it can't be done unfortunately. It is because the near_to and to arguments.

I need to think about this, but I think I've done enough for the day. I do not feel like starting anything else right now.

...No wait, I actually know how to fix this. I just need to rebuild the module. Be right back.