I think this will do it for without.
I seem to be getting out of later and later. Today I got up at 11am and wound up doing chores for nearly two hours. I had a really late start, so I will put in some overtime.

Right now I've put in the new module syntax in the parser and the pevaller, but I need to test it. Let me give it a shot.