Done putting in the `without`. Once again I've increased the usefulness of modules.
In loop6 and hacker_rank_2, these latest changes also caused two methods to be eliminated which just shows that careless passing of state can cause code to be duplicated.

At any rate, let pick a new problem and work on it.

I need to focus. Let me just do this for a while longer.

Oh, yeah, I should ask around on a competitive programming sub if there is a site that supports the latest version of Mono with F# and longer submission sizes.