https://www.hackerrank.com/challenges/botclean
Ok, I will admit that this got me thinking a while.

Really, it is only just not that I noticed that the score is (200 – number_of_moves_made) / 40 and the board is only 25 tiles.

If I was serious, what I would do is start with nearest greedy search and then add monte carlo with some randomization of moves and select the best ones. That is the way to crack this.