I finally passed the submission. The actual solution is maybe 4 lines, but damn does it require some scaffolding.
Well, I did well. This experience made me realize that I am really not even close to being done with patten matching. I am going to have to put in a bunch of optimizations, and maybe build it it in for degenerate unions.

Forget that for now.

What I am going to do from here is make it work for recursive datatypes. Right now it would diverge. I need to put in some extra stuff into the language for this, but it should work.

Let me take a little breather and then I will do this.