The polymorphic equality now works on recursive datatypes, but there is a strange code generation thing that needs investigating.
                let (var_12: bool) = method_18((var_8: Rec0), (var_11: Rec0))
                if var_12 then
                    true
                else
                    false

I am honestly not sure why code like this is geting generated for lists. I can easily add an optimization to take care of it and I have, but just doing it that way is irresponsible.

Before I do it, I want to understand why exactly this is happening because sure as hell I can't imagine this.

Let me investigate a hunch that I have.