https://git_hub.com/Microsoft/visualfsharp/issues/3723
I spent the last hour or so making the above post. Hopefully I will get some better replies there.

That took a bit out of me. So, what is next on the list?

Well, what I would want to do is retouch the problem I did yesterday. After that I will do another Game Theory problem or two while I wait for feedback. I'll give it a few days and then scavenge around if nothing good comes from it.

Also I should just probably do a dictionary on my own based on the hash consing table that I did. I will need to do that anyway to optimize the caching of nodes. But let me leave that for later as it is not that important right now.

Well, let me do some work for a bit.