"Well, I had to think about it for a while. I know how to do it. What I need to do here is simply calculate the number of moves remaining and the winning player should fall out of that.
Let me first start off by calculating all the number of evenly divisible numbers for all numbers [2..1m]."

"This is a difficult problem. I mean, I know that reducing the number of even divisible numbers could be done to number of choices, but in order to solve this problem, I would need to deliver a proof of the condensed pattern. This is not trivial and I would have to simluate it to make sure.

For the number of choice thing, since every number can be prime factored, what I would be doing by the even divisibility thing is simply counting primes. That can be done up to 1m no problem.

Then once you have that, calculating the winner for a single tower is no problem, but for multiple ones...

Erm, I have no idea. Let me check out the forum.

<snip>

Oh, lol, I should not have looked. But condesing it to this point is the skill.

I knew it would be some pattern like that.

Just forget it. I've been spoilered. I need something different from this. Or as a project, I need to figure out how to derive the proof of the above through dynamic programming."

In the end I submitted the solution I found in the discussion forums and saw the editorial. Now I feel like an idiot. Mirroing the moves from player 1 never occured to me.

Crap.

And I had such beautiful ideas for constructing proofs though iteration.

Erm, let me just forget this one and move to the next.