I got spoilered on the Nim problem as well.
Yeah, there is no way I could have come up with the Xor thing. But it did make me realize that I still haven't put in bitwise operators in the language.

The same goes for shift operators. I think it is time I get on with this. After that I will do the Nim problem.