Now Misere Nim passes, but I am honestly tired of these kinds of problems. They are not giving me the kind of exercise that I want.
Save The Princess was really great in that regard since it forced me to evolve the language significantly, but this...not so much.

Right now I am looking at other problems. I'll get back once I've decided.

Now that I've arrived at this point, I am starting to feel the eagerness to get on with making the ML library. I am definitely feeling confident in the language.