I've decided to do some problems from the Recursion and Search sections. I'll start with Hackerland Radio Transmitters.
But that will be left for tomorrow. I do not feel like it anymore.

Today I did enough, like 4 problems in total. Today was possibly the first time I've been able to do several problems in a sequence without discovering some bug in the language or having to upgrade the library.

Streangely enough, despite this being my long awaited 'vacation', I hate it. Now that I feel safer in the langauge so has my ambition grown. Let me just do a day or two more of this and then I start work on the Cuda backend or .NET integration, depending on whether I get the answer to my question.