I am focusing.
val it : MethodInfo =
  T[] SortWith[T](Microsoft.FSharp.Core.FSharpFunc`2[T,Microsoft.FSharp.Core.FSharpFunc`2[T,System.Int32]], T[])

I am thinking about what I should do about this. I should definitely put in MakeGenericMethod and the conversions from closure to FsharpFunc, but that is all.

Using these methods that take in F# lambdas would be simply too messy in Spiral.

I'd be better off writing my own sorting routines. Also I would be better off, in looking into the source for .NET's sort and copying that. Ah, who am I kidding. I never look at other people's code.

Before I do any changes to .NET interop, I want to receive full feedback from Mathias in the thread that I opened.

...

So, it is time to decide what to do?

Well, the first consideration is that writing sorting routines would take me a day or two. They are not something I'd want to do carelessly. I'd want to thoroughly test them out first which would take time.

So I do not want to do it right now.

Interfacing with F# libraries is on the table, but as I said, I want to give it some time before I commit to anything here.

Also I want to change the array representation on the .NET side of things.

So I do not want to do it right now.

Doing these mental exercises is great for motivation and for putting things that I need to do later in my mind.

But for now, how about I do a problem or two that do not require sorting?