Before I start, let me just say that I did it. I was feeling stressed out yesterday over having to skip a problem due to lack of access to sorting routines, but now I have it.
I figure out that the way to put in metadata into `System.Type` is to use the WeakConditionalTable rather mess with inheritance. Now, I am not 100% sure that it will work, that is what I am going to find out right now, but it should work.

The truth is, I should have thought of this weeks ago back when I was doing research on hash consing, but once I found out the table arranged keys by reference I became convinced that it is useless and never touched it again.

But if I think about it more closely, the way it should be doing substitution for the System.Type should preserve the references that come in. So there is a path to making that work.

I also got some alternatives to using the `System.Reflection` from Matthias Dittrich, but I am not interested in dealing with that. All I want to do is resolve this issue in the most elegant way possible.

Let me test this.