Matthias warned me again against using the `System.Type`. I believe him. I'll still go with what I have now for the near term, but this whole picture is shaping up like it will be a large pain in the ass down the read.
Honestly, I do have a responsiblity towards ensuring language integration, but it is not a high priority past putting in tuples and term level functions. Access to .NET is only important to me up to a point. I'll have to interop with C++ and Futhark in the later stages anyway and maybe even some other things so I do not want to commit too heavilty to this here.

If I get complaints about it, I will put in extra effort in this direction.

Right now instead of writing this post I should be diving into F#'s source, but even with VS 2017 I can't open the solution for some reason. I thought for sure it would be because I was using VS 2015, and I did manage to open it a year ago, but right now it fails.

It is done, let me give it a shot.