Done with the F# compiler source dive.
Since things are not going my way, I've been giving way into lethargy and pessimism, but am fighting against that. Really, given the context, Matthias advice is definitely right.

And I know he is right, because looking back, I do remember reading the reflection can build new types at runtime.

Why am I so obsessed with doing things simply? Why not build a struct or a class from scratch to hold all those things? Why did that never occur to me?

It just never even occured to me to even aim for that.

If it did then...

Let me continue thinking along those lines. It seems that I know what my next exercise will be - try to construct a class with a bunch of fields in Cecil.

I have a bunch of other things in mind that I need to do as well.

"I should add the PackedStack record. That way I will be able to move tuples across the boundary without having to split them up. I have a really good thing going on with that particular feature.

Also I should heapify the multidimensional arrays and not store their offsets. Instead I should be calculating that on the fly. I should also heapify the queue.

Also I definitely did not sufficiently test those two features."

This is from my journal. I need to go back a bit and see if heapfying the Queue and the ArrayN will work. I cannot be sure that those features work as intended just yet.

So for the next week or so this is what my goal will be.

I need to level up my metaprogramming, as nasty as it is in .NET, and I need thoroughly test the features that I have and add the packed struct and maybe mutable records while I am at it.

When I am done with that I will have everything I need for absolute integration and will be able to move past the C++ language boundaries without a care.

With that the stage will be set to make the ML library.

Looking past that, I will have one piece missing - resource management. That means I will have to figure out how garbage collection works as well.

But if I am ever at that point, I will be at near victory.

Eventually, I will complete this language and will be able to move to the next step.