I've move the experimentation to here.
This is such a pain I cannot even begin to describe it. Cecil has barely any documenation and everything is unintuitive.

This whole experience has made me see the whole of .NET in a less positive light and is making me reconsider the value of doing integration negatively.

But I will do it of course. I need to do this even though it is sapping my will to live.

I've discovered the extensions in the Rocks namespace, so now I only need to figure out how to make references. Strangely enough, I cannot find an explanation of what the TypeReference is anywhere nor what the scope interface is supposed to be doing.

This is really the kind of job I would not do without pay.

Let me look for some more resources. I can't even figure out how to import a basic `System.Object` type at the moment.