https://stackoverflow.com/questions/46729188/why-does-the-mono-cecil-helloworld-example-fail-with-an-exception
Seriously, what I am trying to do is supposed to be a lot easier than this. The feeling of trying to learn this is like trying to swim in a lake of mud.

Let me take a break for a bit here.

After that I will figure out how to get a TypeReference directly from a definition.