I think I will call it a day here. I am not too tired right now, but I did enough. I do not feel like starting work on anything else (like a HackerRank problem) so late in the day while I wait for a reply. Doing this kind of work is the closest programming gets to bureaucracy.
I will give this a day or two to figure out. If I cannot move out from a dead end, then I won't concern myself too much with .NET interop.

If that happens, I won't consider having only primitive integration my fault - ultimately, the value of interop is inversely proportional with how hard it is.

I have a responsibility to put in honest to God effort into it, but not the level the official MS compiler put into it. Ultimately, if MS wants more people to work with the platform from the outside, it should build better APIs.