https://github.com/jbevain/cecil/issues/454
I opened this issue to get the author's attention. Just now I also did some optimizations for arithmetic operations when one of the sizes is zero.

Now...

http://www.stochasticlifestyle.com/like-julia-scales-productive-insights-julia-developer/

I was reading this article and I do think Julia is impressive. Especially the GPU focus. I need to keep in mind that now the language can do quite a lot.

What I am going to do from here is rediff the cache, heapify the queue and then clone Mono.Cecil and see what it is made of. I do not feel like doing HackerRank exercises while there are more important things to be doing. I want to resolve the issues on hand first. I will also change the CPS'd function in list to use the Option type.

Let me just do this thing for now.