Ok, I am convinced. Heapification and by extension stackification does indeed work.
The bug I was tracking down was not a bug in the language itself, but an incorrect use of partial application in the module.

I'll pass on redesigning the List functions. It is not that important. I do not feel like doing this.

What I will do now is take a little break to do the chores and dive into Mono.Cecil's source. A honest to God's effort will happen just like I promised to myself.