Ok, it passes. Most likely this is the ideal way to indexing. Good job me.
That having said, in the future, I need to avoid these kinds of refactorings. They are just making me waste my time.

I wanted to say that I've gone thourgh the Mono.Cecil source and it is a horde of bureucracy. It is not to say that it is a bad library, it would be more like to say that .NET has too much crud in it.

Interop would be fine if I had an obvious port to plug the language in, but unless I can get help to do this, it will be dead in the water.

The library is complex enough that unless I rewrote it by myself, there is no way I could figure it out. And there is no waaay I am going to do that. .NET interop is not worth worrying about that much.

To be honest, apart from doing these optimizations which are important, I feel like I've been skirting doing more important stuff.

But this is why initiated the push to getting the interop dealth with 3 days ago rather than wait for when I needed it. I wanted to make a decision earlier rather than later since I did not entire trust .NET to make plugging into it easy, and I see that I was right on the money.

Let me redisign list a bit more and then I will call it a day since I am still just mopping up. By convention I should not dyn the state, instead that should be the user's job.

Tomorrow I will start work on the Cuda backend most likely. I want to have both it and the ML library done by the end of the month if possible. Now wouldn't that be something?

I am sure I can do the backend in less than a week. Actually 2 days should be enough, but making interop work will be harder.