As far as redesigns go, this was entirely painless by my previous standards. What I am going to do tomorrow, rather than start on the Cuda backed is completely rework the extension pattern. As it is now, it is completely useless.
Spiral needs some reflective capabilities and pattern matching is a natural target for this.

I've been thinking about reflection for the past few days. I think that I could put in complete macro capabilities in Spiral, in fact I could take a piece of code and convert it to an AST and do an interpreter for it in the language directly that would evaluate it at compile time.

But that kind of thing would be pretty useless.

But improving pattern matching would be pretty great.