I have a very good idea of how to substitute variables in Cecil now, though that depends whether Cecil does the reasonable thing with generic parameters.
https://www.youtube.com/watch?v=DBVLcgq2Eg0

I just finished this PyTorch video. Just as I was thinking that it might be good to switch to PyTorch, the guy comes out and admits that their JIT cannot handle control flow.

That made me lol.

I am being a bit sinister here, I know. In general, they are going in the right direction, but they won't get where they want if they have to fight the big snake. The friction of having to interop with Python will slow them down immensely.

I was intending to watch the Scala video, but let me go forward with my idea and try the variable substitution scheme that I had in mind.

It should work and if it does, I will have the complete interop plan right there and fully workable.