Nevermind, this is just me being retarded. I forgot that the individual JoinPoints store their own data...
But that having said, why the hell am I storing them in a dictionary. Since I allocate a new Tag for every join points, they will never repeat.

I was an idiot, just not in a way I thought. My brain must have been drunk on memoization. I do remember momoizing everything, and so I did this regardless whether it makes sense to do so. Let me get rid of the dictionary and the tag. That way it will be perfect.

Let me fix this.