This is a failure.
I've made an architectural error. The extremely lethargy I am feeling is a side effect of that, it is essentailly my body telling me this can't be done.

The technique where I pass in a message and pattern match on it in order to capture scope and branch on it is a very useful one, but the difference between here and elsewhere is in that the places I've used it before have all been less than 30 lines and I've only branched once on the message.

Here I would literally have to branch everywhere.

What I am trying to do here is the equivalent of trying to merge parsing and typechecking - it could be done in theory, but I am really much better off not doing it.

There is a huge difference between passing a message into a small function and branching once on it, and having to brunch 100 of times in a huge function with no compiler error checking.

I must have been insane to try doing it like this.

What I will do instead is during the F# codegen phase is collect all the Cuda join points and then pass it in the next stage. This is how I separate parsing from the following stages and this won't be an exception.

I'll leave the master branch alone for a bit until I finish the Cuda codegen. Starting here I will branch from an earlier commit.