With this I am done with the first step. Now that I have this, I can seriously start thinking about Cuda interop.
And that will be what I am going to do. I'll introduce the CudaEntry op and a bunch of other stuff from here on out.

I still need to think about how I want to do this exactly. Probly I will introduce the Cuda context and streams as inbuilt types.

I also need to figure out how to install ManagedCuda as a global package. There is really quite a lot to do and I need to do it one by one.

I'll keep putting in the union types for last.