"1:25pm. Let me finally start.
First let me see if I can load the Cuda context. I will start by making a test for that.

1:35pm. I am still thinking.

2pm. ...I am still thinking.

3:40pm. Ok, I reasoned it mostly all out.

Whatever the case is, I am going to take a different approach than last time. And if it was not obvious I am not going to be able to just write cuda (…). Instead it will have to be more like:
cuda_run {
    stream = main_stream
    dims = 128,1,1
    kernel = cuda
        ...
    }
The trick here is that `cuda` desugars to…
inl gridDimX gridDimY gridDimZ -> ...

Actually figuring this out took me a while. I was thinking how to insert the above variables manually when it finally occurred to me that I should just do this during the parsing stage. Brilliant.

3:45pm. One of the difficulties with Cuda interop is that literally everything requires a context and a stream. For example, even allocating arrays is not something that I can just do off the cuff.

One thing I have to do at join points is convert arrays to Cuda arrays before passing them on.

How can I possibly do that at compile time when I do not have a context at hand? It is impossible. The solution in fact is to pass the context as an argument in order to get past that.

3:50pm. What I will do is gather all the kernels into a string and then inside the Cuda module, compile them into PTX at runtime, and then turn that into a lambda that runs the kernel.

4pm. There is one thing I am missing in the language in order to make things work. It is a bit ironic, but I need the array initializer. I need to be able to write array (a,b,c) and have it compile into [|a;b;c|] on the F# side. That way I will be able to just pass the free variables into a tuple.

4:05pm. There is more. I do not feel comfortable doing the array of tuples to a tuple of arrays transformation automatically as it changes the semantics of the program. But unlike for functions I am tracking them only on an individual basis. What I should be doing instead is adding the record fields to them as well since that will open the door to later applying packed_stack to them.

…

In recordify, I need to allow changing of the annotation type in order to make that work.

4:10pm. Calling Spiral `record`s that was dumb since they really determine layout. I should just rename  Record to Layout and it will make a lot more sense than it does now.

There is a lot to do so let me make a list to keep track of things.

1) Add the TupleToArray Op.
2) Rename Record to Layout.
3) Add Layout annotations to tuples.
4) Make sure that the layout annotation is changeable.
5) Add PackedStack layout which corresponds to sequential fully packed struct layout.
6) Simplify the Fix case.
7) In the Ty definition, merge the HeapFunT and StackFunT and all the types that will be introduced into a single case.

#7 Will be something like: LayedT of Layout *  TupleOrFun.

That would be good. That will work.

4:25pm. Is there anything else I am missing here before I start? There is a bunch of stuff beyond the above of course, but the above are just the preliminaries.

4:30pm. Ah, that renamer_apply_env, I needed that to init the damn renamables. Whops.

Let me abstract this a bit.
    let inline renamer_free_variables_template apply x =
        let {fv=fv} as r = renamables0()
        apply r x |> ignore
        fv

    let env_free_variables env = renamer_free_variables_template renamer_apply_env env
    let typed_expr_free_variables exp = renamer_free_variables_template renamer_apply_typedexpr exp

Let me go with this.
        let memoize_closure arg d x =
            let fv = typed_expr_free_variables arg
            let arg_ty = get_type arg

Let me also change this in memoize closure.

I really need to run the tests after all of this. Really, without them, I could not do a single change on the language at this point."

The above is from my journal. The design work for Cuda interop is not as easy at it seems and there needs to back and forth comunication between the host code and the compiler in order to make it work which is the reason why I spend most of the day thinking.

Let me run the tests to make sure I haven't broken everything and then I will checks some things out.

I need to go to the old Spiral library and determine how the kernel to run is selected. I think it by name, but I need to make sure it is not something else.