"cuda_context.LoadKernelPTX(target_path,kernel_name)
Ah, this is bad. It loads it directly from PTX and returns a class representing that kernel.

I wonder if there is an alternative way of doing this?

There is something called Cumodule. I think that should allow me do do what I want.
    let m = cuda_context.LoadModule(target_path)
    CudaKernel("method13",m,cuda_context)

Yeah, it can be done like this. I can just load the module and the get the kernel from it. Easy peasy.

Good. I can manage this.

https://www.englishforums.com/English/LayedVsLaidCorrectUsage/qvczp/post.htm

it seems that layed is not a word.

I guess LaidT it is."

There are no changes introduced by my recent modifications. Thankfully I caught the thing in the codegen.

The above journal excerpt shows how to load kernels from a module. I never used this before, so I was right to be suspicious.