"9:10am. I am up. Now where was I?
I had a good idea last night.

1) Add the TupleToArray Op.
2) Rename Record to Layout.
3) Add Layout annotations to tuples.
4) Make sure that the layout annotation is changeable.
5) Add PackedStack layout which corresponds to sequential fully packed struct layout.
6) Simplify the Fix case.
7) In the Ty definition, merge the HeapFunT and StackFunT and all the types that will be introduced into a single case.
8) Extend CudaGlobal so it takes a Ty.
~9) Rename `eval_renaming` and `eval_method` and the `memo_x`  functions into the `join_point` series.~

Without #8 the coercion will just be too difficult as the codegen won't know which Spiral type to bind it too. This would force me to hardcode the type of CudaGlobal on F# side. Once I add this on the Cuda side, I can just completely ignore the field.

9# is just more streamlining. Like with Layout, the naming scheme for join point related functions is confusing.

9:20am. Done with #9.

9:30am. https://devblogs.nvidia.com/parallelforall/programming-tensor-cores-cuda-9/

Let me read this and that quest for a bit and then I will start programming. I will start by completing the test for loading ManagedCuda.

9:40am. Done reading the Cuda thing. Though I mostly wandered off in thought than focus on it. Let me make a commit and I will read the quest for a little bit. After that I will start working on the tests.

I think I will wait a few more days and then make the resolution to make my own assembly reader. It won't be too difficult, but it will be time consuming, so I will do it after I am done with the Cuda backend and the first iteration of the ML library."

The above is an excerpt from my journal.

With the CudaGlobal idea I think I have now all of the prerequisites met to make the interop work.

Is there anything else? No, I do not think so.

It is just...well, I am doing things out of order here and that makes me uncomfortable.

Having proper .NET interop would make things a lot easier for me. Before I make the 0.1 release of the language, I'd definitely want full .NET interop and that means making that IL reader. Maybe I should look at some of the defunct .NET languages and see how they do it.

F# is just too big - in AbsIL the main file starts with 2k type declarations and then doubles that in the implementation.

The troublesome thing about the whole interop thing is that I can't gauge how hard making that reader will be until I try it.

It has already passed 3.5 months since my last update and I hoped I would have more to show for it than this.

...Agh, I have no choice. Forget what I wrote about doing the backend and the ML library first.

If doing the reader first is the right thing to do then I should just do it.

If I was the kind of person to skim out then I would have never started work on Spiral in the first place.