https://www.ecma-international.org/publications/files/ECMA-ST/ECMA-335.pdf
The book as far as I can see is quite ismilar to the one published 8 years ago. I do not need to read it right now. What I need more are hard specifications that I can turn into machine instructions so let me go with the above for now.

What I need is the data necessary to parse hard machine specs.

Let me see if the above offers it. It should.