In honor of this effort I will do one thing that just occurred to me.
Spiral is pretty powerful, so to the layouts I can add LayoutFsharp. For tuple that would make them be compiled as Fsharp tuples and I can do the same for records. The only requirement is that they do not have literals or type level things in them.

What that will allow me to do is use F#'s native tuples and records in dictionaries.

I won't let anything go to waste.

I will ensure a better level than what I have now through a different means, in a way that flows naturally.