Magnifique.
Now comes the second part, which is arguably more complicated. The F# version uses named arguments and I am not sure how that plus the optional arguments will be dealt with.

Well, let me give it a shot. If I can make the script for calling system processes in Spiral, that will be another great step forward for the language.