"Now where was I? Yeah, let me make some hack for making delegates and then I will move from there.
Like the TryGetValue and the ref conversion these always trip me up and I have to waste time thinking how to deal with them. The TryGetValue thing is especially annoying since it is not like it would have been hard for them to return a tuple or extend the damn thing.

Also the IL book says the delegates are elegant. My ass they are.

The CIL is pretty much a huge collection of special cases each of which make it more difficult to compile for.

And the way F# handles them is special which means if I follow its lead, I will be compiling not to .NET anymore, but to F# specifically."

Special compiler cases are really annoying.

Well, let me focus just on delegates for now, forget the events.