With simple tuple concat, the compilation goes back to the usual 0.1 it takes to Jit the compiler.
Good, good.

I've really been at it today. I can honestly say that I did a decent chunk of work and did it well.

I think I will call it a day here, lunch should be soon anyway.

I have to think of what to do next anyway. Probably, I will save the Cuda kernels to file and load it directly. If I am lucky, I will be able to start work on calling kernels tomorrow.

1) Add the TupleToArray Op.
2) Rename Record to Layout.
3) Add Layout annotations to tuples.
4) Make sure that the layout annotation is changeable.
5) Add PackedStack layout which corresponds to sequential fully packed struct layout.
6) In the Ty definition, merge the HeapFunT and StackFunT and all the types that will be introduced into a single case.
7) Extend CudaGlobal so it takes a Ty.
8) Extend Layout with the LayoutFsharp type.
9) Make sure that layout can be set on types with any number of free variables.
10) Modify sprintf for so %s accepts StringBuilder as well.
11) Modify sprintf so it can be partially evaluable. Also modify it so it can be partially applied – right now the function uses mutable state which prevents that.
12) Modify .NET calls so that when arguments are all literals and the return type is a primitive, they are called at compile time.
13) Add the .NET overrides for String and Array types.

Besides that, I have this list to go through so it is not like there is a lack of work. But I am getting there.