"let q =     """ \" \n" "     """
Ah, triple quoted strings are literal.

But sprintf with triple quoted strings isn't I think.
let w =
    sprintf
        """
\" \n" " %s
        """

Ah, it is with the exception of that %s.

Ok, I do not need to dump it in a file. Triple quoted strings will do nicely."

I have a plan forming. I am going to change cuda2 so it takes the code via string after all instead of loading it from file. This will simplify things.