No forget it. What I had in mind would not work.
That is it. I am giving up on complete .NET interop. Maybe later I will add F# tuples and records, but that is basically it. It is not a priority.

What I am going to do is add a language suggestion to add Item1..n for tuples. Right now they would just be too much of a pain to destructure normally.

Forget this. I've really been in the swing of things yesterday. The closer I get to finishing Spiral and the ML library, the more my mental state seems to improve, so let me keep at it.