I am distracted right now with some other things dumped on my lap. I'll be done with it soon thankfully.
Right now I am dealing with the join point annoyances.

I want to allow the calling of __global__ kernels within device code and it causing me a bit of trouble with pointers and all.

But I think I have that too.

I have it all. Let me try channeling it.